feat(graph):prepare structure for different data models
